+++ js/main.js
@@ -1,7 +1,5 @@
-import './data.js';
+import { generatePhotos } from './data.js';
+import {renderPictures} from './render-picture.js';
 
-import {generatePhotos} from './data.js';
-
-import './rendering-miniature.js';
-
-generatePhotos();
+const pictures = generatePhotos(25);
+renderPictures(pictures);
